DAPWatchExpression: avoid 'Unexpected null value' error when editing expressions (#1200)
This is a proposed fix for #1199. I add a `null` check to the result of
the `adapt` call, as this adaptation may return `null` if you try to
edit or re-evaluate the expression after the program has finished
running (e.g. between repeated debugging sessions of the same program).

Fixes #1199