Semantic tokens from LS available in the Java editor (#1161)
Based on the latest addition of semantic tokens extension for Java
editor the support is added for having semantic tokens from the language
server in the Java editor.
There is a preference add to turn off semantic tokens support in the
Java editor. The preference is defined but the UI is not and is expected
to come from 3rd party plugins using this functionality (i.e. spring
boot ls plugin) otherwise it's unclear where to fit this preference in
the UI at the moment