Synchronize markers updates (#1283)
Sending the same diagnostics from LS twice for example results in race
conditions in `LSPDiagnosticsToMarkers`, namely the job for updating the
Eclipse resource markers in `updateMarkers(...)` method. The job's
scheduling rule is `null`. There is an expression for computing proper
marker modification scheduling rule but the value of the resultant rule
is `null`. Therefore, marker update jobs can run in parallel and result
in some markers duplicated sometimes.
Fixing this cabbe a different scheduling rule, i.e. the resource but
this might affect the build and other things. The other option is just
sync block for updating the markers within the job body. The "sync
block" solution seemed to work and is included in the PR.