fix: Avoid UI freezes on Open Declaration with slow lang servers (#1376)
Without the PR we currently wait up to 800ms which is for some LS too
short (esp. when requests pile up) so the operation may consistently
fail with:

<img width="518" height="148"
alt="159085032-7b63dc48-ccc6-4310-812b-56bc2456a2f9"
src="https://github.com/user-attachments/assets/d9a82797-d75c-4c3a-a0c8-eaa6306b5a92"
/>

Since the wait happens in the UI thread the timeout cannot be extended
without making the UI freeze longer.

With this PR the UI thread will be blocked for max 200ms (or shorter if
we agree on). If the LS does not respond in that time, a
`DeferredOpenDeclarationHyperlink` instance is returned which
asynchronously opens the link once LS responded. Opening the link will
be dismissed if the the editor was closed in the meantime, the document
modified or the response took longer than 5 seconds.

Addresses #97